hostapd: Respect obss_interval on 40 MHz intolerant disconnect
When an AP is configured for HT operation on 40 MHz but has overlapping
BSS on the secondary channel, it will fall back to 20 MHz only. Also a
40 MHz AP will downgrade to 20 MHz while at least one 40 MHz intolerant
STA is connected.

However, in the current code, on such an AP, we do not check the
obss_interval parameter to be set on the last 40 MHz intolerant STA
disconnect and instead, start a timer that will unconditionnaly and
immediately bring the AP to 40 MHz operation even if it had overlapping
BSS.

A similar case was addressed for the 20/40 MHz co-ex report earlier in
commit d027c7b118fe ("Fix 20/40 MHz co-ex report processing with
obss_interval=0"), but this other case that gets called whenever a STA
entry is removed needs to have similar constraint on starting the timer.

Signed-off-by: Nicolas Escande <nico.escande@gmail.com>