mtk: hostapd: Fix Operating Mode Notification issue in 2GHz
If this patch is not applied, since the driver enables the
Operating Mode Notification feature, hostapd will enable the OMN bit
in the Extended Capabilities Element when constructing beacon.
However, this is not allowed on frequency bands that do not support
ieee80211ac, hence add this patch to fix this issue.

Signed-off-by: MeiChia Chiu <MeiChia.Chiu@mediatek.com>